+++ packages/aai-studio-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # aai-studio-client
 
+## 0.6.18
+
+### Patch Changes
+
+- @alexkroman1/aai@12.0.0
+  - @alexkroman1/aai-ui@12.0.0
+
 ## 0.6.17
 
 ### Patch Changes